create template w/o validation needs validation to check if components from request are compatible

+++ CopacoProject/src/main/java/nl/fontys/s3/copacoproject/domain/CustomProduct.java
@@ -13,8 +13,7 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CustomProduct {
-
-    private Integer customProductId;
+    private long customProductId;
     private Template template;
     private User user;
     private List<Component> componentsIncluded;
